clearing_house: bump sub account id to u16 (#211)

+++ programs/clearing_house/src/lib.rs
@@ -44,7 +44,7 @@ pub mod clearing_house {
 
     pub fn initialize_user(
         ctx: Context<InitializeUser>,
-        sub_account_id: u8,
+        sub_account_id: u16,
         name: [u8; 32],
     ) -> Result<()> {
         handle_initialize_user(ctx, sub_account_id, name)
@@ -164,7 +164,7 @@ pub mod clearing_house {
 
     pub fn update_user_name(
         ctx: Context<UpdateUser>,
-        _sub_account_id: u8,
+        _sub_account_id: u16,
         name: [u8; 32],
     ) -> Result<()> {
         handle_update_user_name(ctx, _sub_account_id, name)
@@ -172,7 +172,7 @@ pub mod clearing_house {
 
     pub fn update_user_custom_margin_ratio(
         ctx: Context<UpdateUser>,
-        _sub_account_id: u8,
+        _sub_account_id: u16,
         margin_ratio: u32,
     ) -> Result<()> {
         handle_update_user_custom_margin_ratio(ctx, _sub_account_id, margin_ratio)
@@ -180,7 +180,7 @@ pub mod clearing_house {
 
     pub fn update_user_delegate(
         ctx: Context<UpdateUser>,
-        _sub_account_id: u8,
+        _sub_account_id: u16,
         delegate: Pubkey,
     ) -> Result<()> {
         handle_update_user_delegate(ctx, _sub_account_id, delegate)
